requirements.txt
fastapi
uvicorn
requests
beautifulsoup4
psycopg2
python-dotenv
scikit-learn